+++ src/components/SearchResultSection.tsx
@@ -2,28 +2,11 @@ import * as Tooltip from "@radix-ui/react-tooltip";
 import type { ResponseDocumentMatch } from "@/lib/common";
 import React, { ReactNode, useState } from "react";
 import { Link } from "./Link";
-import { cn } from "@/lib/utils";
+import { cn, getCleanedMetadata } from "@/lib/utils";
+import { AcrobatIcon } from "./ui/acrobat-icon";
 
 interface SearchResultProps {
 	documentMatch: ResponseDocumentMatch | undefined;
-}
-
-function AcrobatIcon() {
-	return (
-		<svg
-			width="18"
-			height="18"
-			viewBox="0 0 24 24"
-			aria-label="Acrobat PDF icon"
-		>
-			<path
-				fill="currentColor"
-				fill-rule="evenodd"
-				d="M1.041 18.061c-.03.031-.059.063-.087.096A3.456 3.456 0 0 0 3.444 24a3.43 3.43 0 0 0 2.413-.967l.046-.04.051-.046.045-.054c.915-1.093 1.945-2.959 2.83-4.62 1.903-.745 3.83-1.468 5.69-1.948 1.73 1.246 4.046 2.053 6.037 2.053A3.456 3.456 0 0 0 24 14.933a3.456 3.456 0 0 0-3.444-3.444c-1.509 0-3.533.466-5.174.978a17.194 17.194 0 0 1-2.724-3.577l.154-.466c.234-.702.473-1.42.667-2.099.278-.967.499-1.962.499-2.88A3.456 3.456 0 0 0 10.533 0 3.456 3.456 0 0 0 7.09 3.444c0 1.003.275 2.064.646 3.08.327.899.75 1.813 1.18 2.685-.69 2.141-1.45 4.338-2.38 6.23l-.786.302c-.711.272-1.405.537-2.027.8-.96.408-1.874.861-2.593 1.44l-.046.038-.042.042Zm10.47-14.617a.967.967 0 0 0-.978-.977.967.967 0 0 0-.977.977c0 .769.5 2.124 1.107 3.537.455-1.452.848-2.803.848-3.537Zm1.979 10.61a19.587 19.587 0 0 1-2.537-3.223c-.51 1.556-1.051 3.137-1.726 4.66 1.397-.532 2.813-1.04 4.263-1.436Zm8.043.88a.967.967 0 0 0-.977-.978c-.915 0-2.483.357-3.92.763 1.252.695 2.818 1.192 3.92 1.192a.967.967 0 0 0 .977-.978ZM4.117 21.28c.478-.576 1.16-1.721 1.848-2.948-1.418.567-2.665 1.099-3.239 1.543a1.05 1.05 0 0 0-.26.68c0 .548.43.977.978.977a.93.93 0 0 0 .673-.252Z"
-				clip-rule="evenodd"
-			/>
-		</svg>
-	);
 }
 
 function ExpandableTableCell({ content }: { content: string }) {
@@ -36,7 +19,7 @@ function ExpandableTableCell({ content }: { content: string }) {
 		: `${content?.slice(0, 200)}${content.length > 200 ? "..." : ""}`;
 
 	return (
-		<p className="text-sm sm:text-base">
+		<p className="text-sm sm:text-base mb-4">
 			{displayedContent}
 			<br />
 			<button
@@ -95,39 +78,8 @@ type SimilarityDisplayProps = {
 };
 
 function SimilarityDisplay(props: SimilarityDisplayProps): ReactNode {
-	const similarityRouded = Math.floor(props.similarity * 10) / 10;
-	return (
-		<Tooltip.Provider delayDuration={0}>
-			<Tooltip.Root>
-				<Tooltip.Trigger asChild className="group shrink-0">
-					<span className="h-1.5 w-8 rounded-full bg-slate-200 inline-block relative">
-						<span
-							className={cn(
-								"h-1.5 w-8 rounded-full absolute top-0 left-0",
-								"transition-colors cursor-pointer",
-								similarityRouded < 0.5 && "bg-red-500 group-hover:bg-red-700",
-								similarityRouded >= 0.5 &&
-									similarityRouded < 0.8 &&
-									"bg-orange-500 group-hover:bg-orange-700",
-								similarityRouded >= 0.8 &&
-									"bg-blue-700 group-hover:bg-blue-900",
-							)}
-							style={{ width: `${similarityRouded * 100}%` }}
-						/>
-					</span>
-				</Tooltip.Trigger>
-				<Tooltip.Portal>
-					<Tooltip.Content
-						className="px-3 py-2 bg-blue-700 text-white shadow-lg rounded animate-in"
-						sideOffset={5}
-					>
-						{Math.floor(props.similarity * 1000) / 10}% Ähnlichkeit
-						<Tooltip.Arrow className="fill-blue-700" />
-					</Tooltip.Content>
-				</Tooltip.Portal>
-			</Tooltip.Root>
-		</Tooltip.Provider>
-	);
+	const similarityRouded = Math.floor(props.similarity * 1000) / 10;
+	return <span className="text-xs">{similarityRouded}% Relevanz</span>;
 }
 
 export default function SearchResultSection({
@@ -152,9 +104,11 @@ export default function SearchResultSection({
 						>
 							<AcrobatIcon />
 							<span>{pdfName}</span>{" "}
-							<span className="text-slate-600 ml-1">
-								(S. {pages?.join(", ")})
-							</span>
+							{pages && pages.length > 0 && (
+								<span className="text-slate-600 ml-1">
+									(S. {pages?.join(", ")})
+								</span>
+							)}
 						</Link>
 					)}
 				</span>
@@ -175,46 +129,4 @@ export default function SearchResultSection({
 			{tags.length > 0 && <TagsList tags={tags} />}
 		</div>
 	);
-}
-
-function getCleanedMetadata(documentMatch: ResponseDocumentMatch | undefined) {
-	const regDoc = documentMatch?.registered_document;
-	const proDoc = documentMatch?.processed_document_summary_match;
-	const metadata =
-		typeof regDoc?.metadata === "object" && regDoc.metadata !== null
-			? regDoc.metadata
-			: {};
-	const type = regDoc?.source_type;
-	const schriftlicheAnfrageTitle =
-		"Titel" in metadata && Array.isArray(metadata.Titel)
-			? `${metadata.Titel[0]}`
-			: undefined;
-	const hauptAusschussProtokollTitle =
-		"title" in metadata && typeof metadata.title === "string"
-			? `${metadata.title}`
-			: undefined;
-	const title =
-		type === "Hauptausschussprotokoll"
-			? hauptAusschussProtokollTitle
-			: schriftlicheAnfrageTitle;
-	const tags = Array.from(
-		new Set(proDoc?.processed_document_summary?.tags ?? []).values(),
-	);
-
-	const pdfUrl = regDoc?.source_url;
-	const pdfName = pdfUrl?.split("/").slice(-1)[0];
-	const pages = documentMatch?.processed_document_chunk_matches
-		.map((c) => c.processed_document_chunk.page)
-		.sort();
-	const similarity = proDoc?.similarity || 0;
-
-	return {
-		title: title?.replace(/<\/?[^>]+(>|$)/g, " ∙ "),
-		pdfUrl,
-		pdfName,
-		pages,
-		similarity,
-		type,
-		tags,
-	};
 }